+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.4.4-SNAPSHOT
 
+* Support multiple stats adapter per compute. Also enhance
+  StatsCollectionTaskState to accept a stats adapter reference. This
+  will allow to collect stats for specific stats adapter.
+
 ## 0.4.3
 
 * Add support for adding and removing collections from resource services